+++ bson2/utils_test.go
@@ -1,0 +1,9 @@
+package bson2
+
+import (
+	"math/big"
+)
+
+func b(i int) *big.Int {
+	return big.NewInt(int64(i))
+}
